Close #24 - Split in packages current architecture.
