--- ecolyzer/__init__.py
@@ -1,3 +1,0 @@
-from .git_analyzer import GitAnalyzer
-from .git_file import GitFile
-from .postgres import Postgres
