+++ ecolyzer/repository/__init__.py
@@ -1,0 +1,1 @@
+from .git_analyzer import GitAnalyzer
